refactor(suite-common): move `acquireDeviceThunk` to `@suite-common/device`, drop `startDiscovery` param
